+++ matrix-sdk/src/main/java/org/matrix/androidsdk/fragments/MatrixMessageListFragment.java
@@ -70,6 +70,7 @@ import org.matrix.androidsdk.rest.model.User;
 import org.matrix.androidsdk.rest.model.VideoMessage;
 import org.matrix.androidsdk.rest.model.bingrules.BingRule;
 import org.matrix.androidsdk.util.ContentManager;
+import org.matrix.androidsdk.util.EventDisplay;
 import org.matrix.androidsdk.util.JsonUtils;
 
 import java.io.File;
@@ -157,9 +158,9 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
     private boolean mLockFwdPagination = true;
 
     protected ArrayList<Event> mResendingEventsList;
-    private HashMap<String, Timer> mPendingRelaunchTimersByEventId = new HashMap<String, Timer>();
+    private HashMap<String, Timer> mPendingRelaunchTimersByEventId = new HashMap<>();
 
-    private HashMap<String, Object> mBingRulesByEventId = new HashMap<String, Object>();
+    private HashMap<String, Object> mBingRulesByEventId = new HashMap<>();
 
     // scroll to to the dedicated index when the device has been rotated
     private int mFirstVisibleRow = -1;
@@ -457,7 +458,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
      */
     protected void onSearchResponse(final SearchResponse searchResponse, final OnSearchResultListener onSearchResultListener) {
         List<SearchResult> searchResults =  searchResponse.searchCategories.roomEvents.results;
-        ArrayList<MessageRow> messageRows = new ArrayList<MessageRow>(searchResults.size());
+        ArrayList<MessageRow> messageRows = new ArrayList<>(searchResults.size());
 
         for(SearchResult searchResult : searchResults) {
             RoomState roomState = null;
@@ -500,6 +501,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
             try {
                 onSearchResultListener.onSearchSucceed(messageRows.size());
             } catch (Exception e) {
+                Log.e(LOG_TAG, "onSearchResponse failed with " + e.getLocalizedMessage());
             }
         }
     }
@@ -549,6 +551,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
                             try {
                                 onSearchResultListener.onSearchFailed();
                             } catch (Exception e) {
+                                Log.e(LOG_TAG, "onSearchResultListener failed with " + e.getLocalizedMessage());
                             }
                         }
                     }
@@ -798,6 +801,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
             }
 
         } catch (Exception e) {
+            Log.e(LOG_TAG, "uploadFileContent failed with " + e.getLocalizedMessage());
         }
 
         // remove any displayed MessageRow with this URL
@@ -853,6 +857,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
             Bitmap thumb = ThumbnailUtils.createVideoThumbnail(uri.getPath(), MediaStore.Images.Thumbnails.MINI_KIND);
             thumbUrl = getMXMediasCache().saveBitmap(thumb, null);
         } catch (Exception e) {
+            Log.e(LOG_TAG, "getVideoThumbailUrl failed with " + e.getLocalizedMessage());
         }
 
         return thumbUrl;
@@ -905,6 +910,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
             uri = Uri.parse(videoUrl);
             thumbUri = Uri.parse(thumbnailUrl);
         } catch (Exception e) {
+            Log.e(LOG_TAG, "uploadVideoContent failed with " + e.getLocalizedMessage());
         }
 
         // the video message is not defined
@@ -1034,6 +1040,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
                 tmpImageMessage.body = uri.getLastPathSegment();
             }
         } catch (Exception e) {
+            Log.e(LOG_TAG, "uploadImageContent failed with " + e.getLocalizedMessage());
         }
 
         // remove any displayed MessageRow with this URL
@@ -1113,6 +1120,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
                 tmpLocationMessage.body = "Location";
             }
         } catch (Exception e) {
+            Log.e(LOG_TAG, "uploadLocationContent failed with " + e.getLocalizedMessage());
         }
 
         // remove any displayed MessageRow with this URL
@@ -1205,7 +1213,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
         Collection<Event> unsent = mSession.getDataHandler().getStore().getUndeliverableEvents(mRoom.getRoomId());
 
         if ((null != unsent) && (unsent.size() > 0)) {
-            mResendingEventsList =  new ArrayList<Event>(unsent);
+            mResendingEventsList =  new ArrayList<>(unsent);
 
             // reset the timestamp
             for (Event event : mResendingEventsList) {
@@ -1219,7 +1227,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
 
     /**
      * Resend an event.
-     * @param event
+     * @param event the event to resend.
      */
     protected void resend(final Event event) {
         // sanity check
@@ -1818,9 +1826,17 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
                                 messageRow.updateEvent(prunedEvent);
                                 JsonObject content = messageRow.getEvent().getContentAsJsonObject();
 
+                                boolean hasToRemoved = (null == content) || (null == content.entrySet()) || (0 == content.entrySet().size());
+
+                                // test if the event is displayable
+                                if (!hasToRemoved) {
+                                    EventDisplay eventDisplay = new EventDisplay(getActivity(), prunedEvent, roomState);
+                                    hasToRemoved = TextUtils.isEmpty(eventDisplay.getTextualDisplay());
+                                }
+
                                 // event is removed if it has no more content.
-                                if ((null == content) || (null == content.entrySet()) || (0 == content.entrySet().size())) {
-                                    mAdapter.removeEventById(event.getRedacts());
+                                if (hasToRemoved) {
+                                    mAdapter.removeEventById(prunedEvent.eventId);
                                 }
                             }
 
@@ -1896,6 +1912,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
             }
 
         } catch (Exception e) {
+            Log.e(LOG_TAG, "onReceiptEvent failed with " + e.getLocalizedMessage());
         }
 
         if (shouldRefresh) {
@@ -2010,6 +2027,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
                 try {
                     mRoomPreviewDataListener = (RoomPreviewDataListener) getActivity();
                 } catch (ClassCastException e) {
+                    Log.e(LOG_TAG, "getRoomPreviewData failed with " + e.getLocalizedMessage());
                 }
             }
 
@@ -2027,47 +2045,59 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
     }
 
     /***  MessageAdapter listener  ***/
+    @Override
     public void onRowClick(int position) {
     }
-
+    @Override
     public boolean onRowLongClick(int position) {
         return false;
     }
 
+    @Override
     public void onContentClick(int position) {
     }
 
+    @Override
     public boolean onContentLongClick(int position) {
         return false;
     }
 
+    @Override
     public void onAvatarClick(String userId) {
     }
 
+    @Override
     public boolean onAvatarLongClick(String userId) {
         return false;
     }
 
+    @Override
     public void onSenderNameClick(String userId, String displayName) {
     }
 
+    @Override
     public void onMediaDownloaded(int position) {
     }
 
+    @Override
     public void onReadReceiptClick(String eventId, String userId, ReceiptData receipt) {
     }
 
+    @Override
     public boolean onReadReceiptLongClick(String eventId, String userId, ReceiptData receipt) {
         return false;
     }
 
+    @Override
     public void onMoreReadReceiptClick(String eventId) {
     }
 
+    @Override
     public boolean onMoreReadReceiptLongClick(String eventId) {
         return false;
     }
 
+    @Override
     public void onURLClick(Uri uri) {
         if (null != uri) {
             Intent intent = new Intent(Intent.ACTION_VIEW, uri);
@@ -2076,6 +2106,7 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
         }
     }
 
+    @Override
     public boolean shouldHighlightEvent(Event event) {
         String eventId = event.eventId;
 
@@ -2101,6 +2132,22 @@ public class MatrixMessageListFragment extends Fragment implements MatrixMessage
         }
 
         return res;
+    }
+
+    @Override
+    public void onMatrixUserIdClick(String userId) {
+    }
+
+    @Override
+    public void onRoomAliasClick(String roomAlias) {
+    }
+
+    @Override
+    public void onRoomIdClick(String roomId) {
+    }
+
+    @Override
+    public void onMessageIdClick(String messageId) {
     }
 
     // thumbnails management